requirements.txt
discord.py==2.4.0
aiosqlite==0.20.0
httpx==0.28.1
apscheduler==3.11.0
python-dotenv==1.0.1
openai==1.60.0